% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.R
\name{bn}
\alias{bn}
\title{Getting suitable genetic IVs through Bayesian network learning}
\usage{
bn(
  df,
  snp,
  exposureName,
  bn_method = "hr",
  cutoff = 0.7,
  repeats = 100,
  nsam = 500
)
}
\arguments{
\item{df}{a data frame which contains data of SNPs and specified exposure.}

\item{snp}{a vector of string belonging to colnames of df, which is the name of SNPs included in BN structure learning.}

\item{exposureName}{a string which is a colname of df corresponding to the exposure studied.}

\item{bn_method}{method for BN structure learning. Possible values are the function name of structure learning algorithm implemented in bnlearn. Default is "hc".}

\item{cutoff}{a numeric between 0 to 1. Those SNPs with score larger than "cutoff" will be chosen as IVs. Default is 0.7.}

\item{repeats}{an integer standing for the times of bootstraping. Default is 100.}

\item{nsam}{an integer standing for the sample size for bootstraping sampling. Default is 500.}
}
\value{
a list containing:
  selectsnp: a vector of string containing the colnames of df corresponding to
  selected SNPs.
  dfscore: a data frame containing the score calculated for each SNP.
}
\description{
is used to get the suitable SNPs as instrumental variables (IVs) of specified exposure by Bayesian network (BN) structure learning.
}
\examples{


}
