% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnmr.R
\name{bnmr}
\alias{bnmr}
\title{Bayesian Mendelian randomization based on Bayesian network}
\usage{
bnmr(
  df,
  snp,
  exposureName,
  outcomeName,
  bn_method = "hr",
  mr_model = "linear",
  cutoff = 0.7,
  repeats = 100,
  nsam = 500,
  init = "median",
  n.iter = 500
)
}
\arguments{
\item{df}{a data frame which contains data of SNPs, specified exposure and outcome.}

\item{snp}{a vector of string belonging to colnames of df, which is the name of SNPs included in BN structure learning.}

\item{exposureName}{a string which is a colname of df corresponding to the exposure studied.}

\item{outcomeName}{a string which is a colname of df corresponding to the outcome studied.}

\item{bn_method}{method for BN structure learning. Possible values are the function name of structure learning algorithm implemented in bnlearn. Default is "hc".}

\item{mr_model}{model for MR. Possible values are "linear" or "logit". Default is "linear".}

\item{cutoff}{a numeric between 0 to 1. Those SNPs with score larger than "cutoff" will be chosen as IVs. Default is 0.7.}

\item{repeats}{an integer standing for the times of bootstraping. Default is 100.}

\item{nsam}{an integer standing for the sample size for bootstraping sampling. Default is 500.}

\item{init}{the init value of theta for MCMC estimation. It can be a specific numeric or a string of "median", "egger" and "ivw", which means the initial value of the iteration will be calculated automatically by the above method. Default is "median".}

\item{n.iter}{an integer standing for the number of iterations. Default is 500.}
}
\value{
a list containing:
  selectsnp: a vector of string containing the colnames of df corresponding to selected SNPs.
  dfscore: a data frame containing the score calculated for each SNP.
  thetaList: a vector cantaining the result of MCMC sampling of the causal parameter we want to estimate.
  mean: the mean estimate of the causal parameter.
  se: the SE of the estimation.
  sd: the SD of the estimation.
  lower: the lower boundary of the 95% CI of the causal estimation.
  upper: the upper boundary of the 95% CI of the causal estimation.
  Rhat: a indicator to measure the convergence (at convergence, Rhat <= 1.05).
}
\description{
to get the causal estimation between exposure and outcome in the presence of pleiotropy by learning Bayesian network structure on GWAS data to get suitable genetic IVs for adaptive Mendelian randomization.
}
\examples{


}
