% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exh_abf.R
\name{exh.abf}
\alias{exh.abf}
\title{ABF calculation for GWAS meta-analysis through Subset-Exhaustive}
\usage{
exh.abf(
  betas,
  ses,
  prior.sigma,
  prior.cor = "indep",
  prior.rho = NA,
  cryptic.cor = NA,
  log = FALSE,
  log10 = FALSE,
  study.names = NULL,
  na.rm = FALSE,
  tolerance = 0
)
}
\arguments{
\item{betas}{a numeric vector of observed effect sizes of a single SNP in a set of studies. Each element of the vector is assumed to correspond to a study.}

\item{ses}{a numeric vector of standard errors corresponding to those in betas. It should have the same length as betas.}

\item{prior.sigma}{the prior on true effect sizes for each SNP in each study. It can be a flat value, set for each study (i.e. a vector whose length is equal to the number of studies in the meta-analysis) or set for each study and SNP (i.e. a matrix of same dimension as betas).}

\item{prior.cor}{a square matrix whose row and column numbers are the same as the number of studies. Its elements are the pairwise correlations between true effect sizes of the studies. It can take values "indep" (independent effects), "fixed" (fixed effects), "correlated" (correlated effects, which requires the prior.rho parameter to be set), as well as individual matrices. If betas and ses are matrices, the same prior.cor will be applied to every row (representing every SNP).}

\item{prior.rho}{either a single value or the upper triangle of a correlation matrix for the prior.cor matrix when it is set to "correlated". If this value is set, but prior.cor is not set to "correlated", this parameter will be ignored.}

\item{cryptic.cor}{a square matrix whose row and coumn numbers are the same as the number of studies. If the studies in the meta-analysis are not independent of each other, it may be necessary to set this parameter so that the covariance in null effects is accounted for.}

\item{log}{sets whether the answer should be given in log space.}

\item{log10}{sets whether the answer should be given in log10 space.}

\item{study.names}{if set, the output will label the columns with the study names.}

\item{na.rm}{if there are NAs in the data, these are removed and the calculation is performed with the remaining data. This happens regardless of how this parameter is set. By default, the output will include a column of NAs for the study with the missing data. Changing this parameter to TRUE removes this column.}

\item{tolerance}{for the ABF calculation, this can be lowered (or raised, if necessary) if the answers are not what was expected. Should probably never be altered, but is there in case it is needed. Default is 1e-1000.}
}
\value{
a data frame containing ABF calculated in all subsets.
}
\description{
Subset-Exhaustive to get the ABFs in all subsets for a single SNP and a phenotype in GWAS meta-analysis.
}
