% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr}
\alias{mr}
\title{Causal estimation by adaptive Bayesian Mendelian randomization with pleiotropy}
\usage{
mr(
  df,
  selectsnp,
  exposureName,
  outcomeName,
  mr_model = "linear",
  init = "median",
  n.iter = 500
)
}
\arguments{
\item{df}{a data frame which contains data of IVs, specified exposure and outcome.}

\item{selectsnp}{a vector of string containing the colnames of df corresponding to the IV used in MR.}

\item{exposureName}{a string which is a colname of df corresponding to the exposure studied.}

\item{outcomeName}{a string which is a colname of df corresponding to the outcome studied.}

\item{mr_model}{model for MR. Possible values are "linear" or "logit". Default is "linear".}

\item{init}{the init value of theta for MCMC estimation. It can be a specific numeric or a string of "median", "egger" and "ivw", which means the initial value of the iteration will be calculated automatically by the above method. Default is "median".}

\item{n.iter}{an integer standing for the number of iterations. Default is 500.}
}
\value{
a list containing:
  thetaList: a vector cantaining the result of MCMC sampling of the causal parameter we want to estimate.
  mean: the mean estimate of the causal parameter.
  se: the SE of the estimation.
  sd: the SD of the estimation.
  lower: the lower boundary of the 95% CI of the causal estimation.
  upper: the upper boundary of the 95% CI of the causal estimation.
  Rhat: a indicator to measure the convergence (at convergence, Rhat <= 1.05).
}
\description{
is used to get the causal estimation between exposure and outcome through Bayesian adaptive Mendelian randomization with pleiotropy after IV selection.
}
\examples{


}
