% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_shotgun_abf.R
\name{multi_shotgun_abf}
\alias{multi_shotgun_abf}
\title{Title ABF calculation across multiple variants for GWAS meta-analysis through shotgun stochastic search}
\usage{
multi_shotgun_abf(
  df,
  vname = 1,
  vbetas = seq(2, ncol(df), 2),
  vses = seq(3, ncol(df), 2),
  prior.sigma = 0.5,
  prior.cor = "indep",
  prior.rho = NA,
  cryptic.cor = NA,
  log = FALSE,
  log10 = FALSE,
  na.rm = FALSE,
  tolerance = 0,
  n.iter = 50,
  B = 5
)
}
\arguments{
\item{df}{a data frame containing the names of variants, the effect sizes and standard errors for each variant in each study.}

\item{vname}{the index of column of variant name in df. Default is 1.}

\item{vbetas}{a vectors to represent the index of columns of those containing the effect sizes in each study, default is the column 2,4,6,etc.}

\item{vses}{a vectors to represent the index of columns of those containing the corresponding standard errors in each study, default is the column 3,5,7,etc. The length of vses should be the same as the length of vbetas.}

\item{prior.sigma}{the prior on true effect sizes for each SNP in each study. It can be a flat value, set for each study (i.e. a vector whose length is equal to the number of studies in the meta-analysis) or set for each study and SNP (i.e. a matrix of same dimension as betas).}

\item{prior.cor}{a square matrix whose row and column numbers are the same as the number of studies. Its elements are the pairwise correlations between true effect sizes of the studies. It can take values "indep" (independent effects), "fixed" (fixed effects), "correlated" (correlated effects, which requires the prior.rho parameter to be set), as well as individual matrices. If betas and ses are matrices, the same prior.cor will be applied to every row (representing every SNP).}

\item{prior.rho}{either a single value or the upper triangle of a correlation matrix for the prior.cor matrix when it is set to "correlated". If this value is set, but prior.cor is not set to "correlated", this parameter will be ignored.}

\item{cryptic.cor}{a square matrix whose row and coumn numbers are the same as the number of studies. If the studies in the meta-analysis are not independent of each other, it may be necessary to set this parameter so that the covariance in null effects is accounted for.}

\item{log}{sets whether the answer should be given in log space.}

\item{log10}{sets whether the answer should be given in log10 space.}

\item{na.rm}{if there are NAs in the data, these are removed and the calculation is performed with the remaining data. This happens regardless of how this parameter is set. By default, the output will include a column of NAs for the study with the missing data. Changing this parameter to TRUE removes this column.}

\item{tolerance}{for the ABF calculation, this can be lowered (or raised, if necessary) if the answers are not what was expected. Should probably never be altered, but is there in case it is needed. Default is 1e-1000.}

\item{n.iter}{the number of iteration for MCMC. Default is 50.}

\item{B}{the largest number of subsets involved in the optimal set. Default is 5.}
}
\value{
a data frame containing:

SNP: Variant Name
ABF: the final ABF value calculated for each variant
model: a string of 0-1 to represents the subset model selected to calculate.
n_studies: the number of studies involved for each variant. If needClean is FALSE, the value will be the same for all variants.
studies_involved: a string of 0-1 to represents the studies involved for each variant. 1 represents the corresponding study is involved.
}
\description{
using shotgun stochastic search algorithm to quickly get the optimal ABF for multiple SNPs and a phenotype in GWAS meta-analysis.
}
\examples{
library(GWASmeta)
data(multi)
re <- multi_shotgun_abf(multi)
}
